backend/requirements.txt
Flask
pandas
scikit-learn
google-generativeai>=0.8.0
python-dotenv
werkzeug
py-googletrans
Pillow
PyMuPDF
requests
beautifulsoup4
Flask-Cors
httpx==0.26.0
gunicorn
